+++ packages/components/organisms/f-offers/src/tenants/es-ES.js
@@ -1,4 +1,38 @@
-export default {
+const messages = {
     locale: 'es-ES',
-    text: 'I am a VOffers Component (ES)'
+
+    noOffersFound: {
+        title: '¿No encuentras nada?',
+        subtitle: 'Puede que todavía haya restaurantes cerca de ti con ofertas o descuentos.'
+    },
+
+    authenticated: {
+        header: {
+            title: 'Te mereces un capricho',
+            subtitle: 'Estas ofertas son personalizadas para ti, demasiado buenas para perdérselas.'
+        },
+        voucherCodeModal: {
+            title: 'Código de oferta copiado',
+            body: 'Introduce el código en el apartado "Tengo un código descuento" a la hora de hacer el pago',
+            confirmButton: 'Pide ahora',
+            dismissButton: 'Todavía no, gracias'
+        }
+    },
+
+    unauthenticated: {
+        header: {
+            title: 'Los mejores restaurantes a domicilio',
+            subtitle: 'Haz tu pedido a domicilio o para recoger en la mejor selección de restaurantes de España'
+        },
+        title: 'Inicia sesión para ver tus ofertas',
+        subtitle: 'Sabrosos descuentos pueden estar esperándote.',
+        loginButton: 'Inicia sesión',
+        signUpButton: 'Inscríbete'
+    },
+
+    termsAndConditionsLinkText: 'Términos y condiciones aquí',
+    termsAndConditionsLinkUrl: '/info/terminos-y-condiciones#ii.terminos-y-condiciones-de-los-cupones-promocionales-just-eat'
+};
+export default {
+    messages
 };