+++ packages/services/f-analytics/src/services/analytics.mapper.js
@@ -10,27 +10,6 @@ import {
 } from '../constants';
 
 /**
- * Returns the current display width name (if clientside).
- *
- * @return {string} The display width name
- */
-const getDisplaySize = () => {
-    if (typeof (window) !== 'undefined') {
-        const width = window.innerWidth;
-
-        // Use fozzie for breakpoints?
-        if (width > 1280) return 'full-size';
-        else if (width > 1025) return 'huge';
-        else if (width > 768) return 'wide';
-        else if (width > 414) return 'mid';
-
-        return 'narrow';
-    }
-
-    return undefined;
-};
-
-/**
  * Returns the current orientation name (if clientside).
  *
  * @return {string} The orientation name
@@ -72,29 +51,12 @@ const mapAnonymousUserId = req => {
 };
 
 /**
- * Maps the user agent string (if present) to the PlatformData 'userAgent' field
- * Note: if executed clientside then the value will be read from the `window.navigator`
- * otherside it is read from the 'user-agent' header.
- *
- * @param {object} req - The `request` context
- * @return {string} userAgentString
- */
-const mapUserAgent = req => {
-    let userAgentString;
-    if (typeof (window) !== 'undefined' && window.navigator) {
-        userAgentString = window.navigator.userAgent;
-    } else if (req && req.headers) {
-        userAgentString = req.headers['user-agent'];
-    }
-    return userAgentString;
-};
-
-/**
  * Maps various environment variables (if present); that are only available when executed
  * serverside, to the PlatformData.
  * Also maps the user percentage experiment value (if present) to the PlatformData; again
  * only available serverside due to it's protection.
- * Note: this is stored until the rest of the PlatformData is collated and sent clientside.
+ * Note 1: this is stored until the rest of the PlatformData is collated and sent clientside.
+ * Note 2: dot notation on the env vars does not bundle well so using valid alternative [] approach.
  *
  * @param {object} platformData - A reference to the current PlatformData instance
  * @param {object} req - The `request` context
@@ -103,13 +65,14 @@ const mapUserAgent = req => {
 export const mapServerSidePlatformData = ({ platformData, req } = {}) => {
     const userPercent = getCookie('je-user_percentage', req);
 
+    /* eslint-disable dot-notation */
     const mappedData = {
         ...platformData,
         jeUserPercentage: userPercent || platformData.jeUserPercentage,
-        environment: process.env.justEatEnvironment || platformData.environment,
-        version: process.env.FEATURE_VERSION || platformData.version,
-        instancePosition: process.env.INSTANCE_POSITION || platformData.instancePosition,
-        isPilot: process.env.IS_PILOT || platformData.isPilot
+        environment: process.env['justEatEnvironment'] || platformData.environment,
+        version: process.env['FEATURE_VERSION'] || platformData.version,
+        instancePosition: process.env['INSTANCE_POSITION'] || platformData.instancePosition,
+        isPilot: process.env['IS_PILOT'] || platformData.isPilot
     };
 
     return mappedData;
@@ -121,21 +84,17 @@ export const mapServerSidePlatformData = ({ platformData, req } = {}) => {
  * @param {object} platformData - A reference to the current PlatformData instance
  * @param {string} featureName - The name of the feature
  * @param {string} locale - The current locale
- * @param {object} req - The `request` context
  * @return {object} new platformData object
  */
 export const mapPlatformData = ({
-    platformData, featureName, locale, req
+    platformData, featureName, locale
 } = {}) => {
-    const userAgent = mapUserAgent(req);
     const mappedPlatformData = {
         ...platformData,
-        name: featureName,
-        userAgent
+        name: featureName
     };
     mappedPlatformData.appType = DEFAULT_APP_TYPE;
     mappedPlatformData.applicationId = DEFAULT_APP_ID;
-    mappedPlatformData.branding = COUNTRY_INFO[locale].brand;
     mappedPlatformData.country = COUNTRY_INFO[locale].country;
     mappedPlatformData.language = COUNTRY_INFO[locale].language;
     mappedPlatformData.currency = COUNTRY_INFO[locale].currency;
@@ -185,33 +144,25 @@ export const mapUserData = ({ userData, authToken, req } = {}) => {
  * Maps various static/computed variables to the PageData.
  *
  * @param {object} pageData - A reference to the current PageData instance
- * @param {string} featureName - The name of the feature
  * @param {string} pageName - The name of the page
- * @param {string} requestId - The current request Id
  * @param {number} httpStatusCode - The httpStatusCode (only supplied when 200 needs to be overriden)
  * @param {object} req - The `request` context
  * @return {object} new pageData object
  */
 export const mapPageData = ({
     pageData,
-    featureName,
     pageName,
-    requestId,
     httpStatusCode,
     req
 } = {}) => {
     const conversationId = getCookie('x-je-conversation', req);
-    const displaySize = getDisplaySize();
     const orientation = getOrientation();
 
     const mappedPageData = {
         ...pageData,
-        group: featureName,
         name: pageName || pageData.pageName,
         conversationId: conversationId || pageData.conversationId,
-        requestId: requestId || pageData.requestId,
         httpStatusCode: httpStatusCode || pageData.httpStatusCode,
-        display: displaySize || pageData.display,
         orientation: orientation || pageData.orientation
     };