+++ packages/services/f-analytics/src/store/analytics.module.js
@@ -13,8 +13,6 @@ export default {
             name: undefined,
             appType: undefined,
             applicationId: undefined,
-            userAgent: undefined,
-            branding: undefined,
             country: undefined,
             language: undefined,
             jeUserPercentage: undefined,
@@ -32,13 +30,10 @@ export default {
             signupDate: undefined
         },
         pageData: {
-            group: undefined,
             name: undefined,
             httpStatusCode: 200,
             conversationId: undefined,
-            requestId: undefined,
-            orientation: undefined,
-            display: undefined
+            orientation: undefined
         },
         events: []
     }),