+++ packages/services/f-analytics/src/tests/helpers/setup.js
@@ -8,8 +8,6 @@ const defaultState = {
         name: undefined,
         appType: undefined,
         applicationId: undefined,
-        userAgent: undefined,
-        branding: undefined,
         country: undefined,
         language: undefined,
         jeUserPercentage: undefined,
@@ -27,13 +25,10 @@ const defaultState = {
         signupDate: undefined
     },
     pageData: {
-        group: undefined,
         name: undefined,
         httpStatusCode: 200,
         conversationId: undefined,
-        requestId: undefined,
-        orientation: undefined,
-        display: undefined
+        orientation: undefined
     },
     events: []
 };
@@ -44,8 +39,6 @@ const modifiedState = {
         name: 'test-name',
         appType: 'test-appType',
         applicationId: 9,
-        userAgent: 'test-userAgent',
-        branding: 'test-branding',
         country: 'zu',
         language: 'ze',
         jeUserPercentage: 88,
@@ -64,12 +57,9 @@ const modifiedState = {
     },
     pageData: {
         name: 'test-name',
-        group: 'test-group',
         httpStatusCode: 200,
         conversationId: '460cc3a8-83f7-4e80-bb46-c8a69967f249',
-        requestId: '6cbe6509-9122-4e66-a90a-cc483c34282e',
-        orientation: 'Landscape',
-        display: 'wide'
+        orientation: 'Landscape'
     }
 };