+++ packages/services/f-analytics/CHANGELOG.md
@@ -3,6 +3,24 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+
+v0.16.0
+------------------------------
+*September 24, 2021*
+
+### Changed
+- Removed the redundant fields from  pushPlatformData() & pushPageData().
+
+
+v0.15.0
+------------------------------
+*September 21, 2021*
+
+### Changed
+- Always attempt to register the store but always preserve state, if present.
+- Removed the redundant `groupName` prop from pushPageData().
+
+
 v0.14.0
 ------------------------------
 *September 10, 2021*